sources/requirements.txt
Flask>=2.0.0
python-docx==0.8.11
pandas==1.3.5 #1.5.3
numpy==1.21.6 #1.23.5
Pillow==9.5.0
requests==2.28.1
Werkzeug>=2.0.0